+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * A `preview` property to the `Article` object.
+* Undo API for Outbox items.
+
 
 ## [5.2.0] - 2025-02-13